[Fix] Modificato delete post

+++ posts/delete.php
@@ -1,4 +1,6 @@
 <?php 
+session_start();
+
 include "../db_connection.php";
 
 // Verifica se è stato inviato un parametro "id" tramite GET
@@ -6,9 +8,9 @@ if(isset($_GET['id']) && !empty($_GET['id'])) {
     $id = $_GET['id'];
 
     // Preparazione e esecuzione della query di eliminazione
-    $sql = "DELETE FROM posts WHERE id = ?";
+    $sql = "DELETE FROM posts WHERE id = ? AND user_id=?";
     $stmt = $conn->prepare($sql);
-    $stmt->bind_param("i", $id);
+    $stmt->bind_param("ii", $id, $_SESSION['user_id']);
 
     include "../error_db.php";
